+++ spec/llm_cost_tracker/ledger/store_spec.rb
@@ -25,7 +25,7 @@ RSpec.describe "ActiveRecord storage integration" do
 
   def track_and_flush(**kwargs)
     event = LlmCostTracker.track(**kwargs)
-    LlmCostTracker.flush!
+    LlmCostTracker::Ingestion::Worker.flush!
     event
   end